+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Acquisition functions are now their own objects
 - `acquisition_function_cls` constructor parameter renamed to `acquisition_function`
 
+### Removed
+- `model_params` attribute from `Surrogate` base class, `GaussianProcessSurrogate` and
+  `CustomONNXSurrogate`
+  
 ### Fixed
 - `n_task_params` now evaluates to 1 if `task_idx == 0`
 - Simulation no longer fails in `ignore` mode when lookup dataframe contains duplicate